# Conflicts:
#	Source/DafnyCore/AST/SystemModuleManager.cs

+++ Source/Scripts/IEncoder.cs
@@ -1,8 +1,10 @@
+using System.Numerics;
+
 namespace Scripts;
 
 public interface IEncoder {
   void WriteNullable(bool isNull);
-  void WriteInt(int value);
+  void WriteInt(BigInteger value);
   void WriteQualifiedName(string name);
   void WriteString(string value);
   void WriteBool(bool value);
